requirements.txt
python-dotenv

asana
langchain
langchain_community